comment, request to chage formatting
added a comment requesting to change the div that holds the written content to be centred and with a max width of 800px or like 80%, currently with longer sentences it streches uneavenly to the right.
elsewise if u can limit the div to not go into the last 1-2 collumns? idk how the collumns work fully, but if u can stop the text from flowing over into the last one or two collumns it would look better imo?